+++ src/Form/Color.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Okipa\LaravelBootstrapComponents\Form;
+
+class Color extends Input
+{
+    /**
+     * The component config key.
+     *
+     * @property string $view
+     */
+    protected $configKey = 'form.color';
+    /**
+     * The input type.
+     *
+     * @property string $type
+     */
+    protected $type = 'color';
+}