+++ tests/Unit/Form/ColorTest.php
@@ -7,67 +7,67 @@ use Okipa\LaravelBootstrapComponents\Form\Input;
 use Okipa\LaravelBootstrapComponents\Test\BootstrapComponentsTestCase;
 use Okipa\LaravelBootstrapComponents\Test\Fakers\UsersFaker;
 
-class TextTest extends BootstrapComponentsTestCase
+class ColorTest extends BootstrapComponentsTestCase
 {
     use UsersFaker;
 
     public function testConfigStructure()
     {
         // components.form
-        $this->assertTrue(array_key_exists('text', config('bootstrap-components.form')));
-        // components.form.text
-        $this->assertTrue(array_key_exists('view', config('bootstrap-components.form.text')));
-        $this->assertTrue(array_key_exists('icon', config('bootstrap-components.form.text')));
-        $this->assertTrue(array_key_exists('legend', config('bootstrap-components.form.text')));
-        $this->assertTrue(array_key_exists('class', config('bootstrap-components.form.text')));
-        $this->assertTrue(array_key_exists('html_attributes', config('bootstrap-components.form.text')));
-        // components.form.text.class
-        $this->assertTrue(array_key_exists('container', config('bootstrap-components.form.text.class')));
-        $this->assertTrue(array_key_exists('component', config('bootstrap-components.form.text.class')));
-        // components.form.text.html_attributes
-        $this->assertTrue(array_key_exists('container', config('bootstrap-components.form.text.html_attributes')));
-        $this->assertTrue(array_key_exists('component', config('bootstrap-components.form.text.html_attributes')));
+        $this->assertTrue(array_key_exists('color', config('bootstrap-components.form')));
+        // components.form.color
+        $this->assertTrue(array_key_exists('view', config('bootstrap-components.form.color')));
+        $this->assertTrue(array_key_exists('icon', config('bootstrap-components.form.color')));
+        $this->assertTrue(array_key_exists('legend', config('bootstrap-components.form.color')));
+        $this->assertTrue(array_key_exists('class', config('bootstrap-components.form.color')));
+        $this->assertTrue(array_key_exists('html_attributes', config('bootstrap-components.form.color')));
+        // components.form.color.class
+        $this->assertTrue(array_key_exists('container', config('bootstrap-components.form.color.class')));
+        $this->assertTrue(array_key_exists('component', config('bootstrap-components.form.color.class')));
+        // components.form.color.html_attributes
+        $this->assertTrue(array_key_exists('container', config('bootstrap-components.form.color.html_attributes')));
+        $this->assertTrue(array_key_exists('component', config('bootstrap-components.form.color.html_attributes')));
     }
 
     public function testExtendsInput()
     {
-        $this->assertEquals(Input::class, get_parent_class(bsText()));
+        $this->assertEquals(Input::class, get_parent_class(bsColor()));
     }
 
     public function testName()
     {
-        $html = bsText()->name('name')->toHtml();
+        $html = bsColor()->name('name')->toHtml();
         $this->assertContains('name="name"', $html);
     }
 
     public function testType()
     {
-        $html = bsText()->name('name')->toHtml();
-        $this->assertContains('type="text"', $html);
+        $html = bsColor()->name('name')->toHtml();
+        $this->assertContains('type="color"', $html);
     }
 
     /**
      * @expectedException \Exception
-     * @expectedExceptionMessage Okipa\LaravelBootstrapComponents\Form\Text : Missing $name property. Please use the
+     * @expectedExceptionMessage Okipa\LaravelBootstrapComponents\Form\Color : Missing $name property. Please use the
      *                           name() method to set a name.
      */
     public function testInputWithoutName()
     {
-        bsText()->toHtml();
+        bsColor()->toHtml();
     }
 
     public function testModelValue()
     {
         $user = $this->createUniqueUser();
-        $html = bsText()->model($user)->name('name')->toHtml();
+        $html = bsColor()->model($user)->name('name')->toHtml();
         $this->assertContains('value="' . $user->name . '"', $html);
     }
 
     public function testConfigIcon()
     {
         $configIcon = 'test-config-icon';
-        config()->set('bootstrap-components.form.text.icon', $configIcon);
-        $html = bsText()->name('name')->toHtml();
+        config()->set('bootstrap-components.form.color.icon', $configIcon);
+        $html = bsColor()->name('name')->toHtml();
         $this->assertContains('<span class="icon input-group-text">' . $configIcon . '</span>', $html);
     }
 
@@ -75,34 +75,34 @@ class TextTest extends BootstrapComponentsTestCase
     {
         $configIcon = 'test-config-icon';
         $customIcon = 'test-custom-icon';
-        config()->set('bootstrap-components.form.text.icon', $configIcon);
-        $html = bsText()->name('name')->icon($customIcon)->toHtml();
+        config()->set('bootstrap-components.form.color.icon', $configIcon);
+        $html = bsColor()->name('name')->icon($customIcon)->toHtml();
         $this->assertContains('<span class="icon input-group-text">' . $customIcon . '</span>', $html);
         $this->assertNotContains('<span class="icon input-group-text">' . $configIcon . '</span>', $html);
     }
 
     public function testNoIcon()
     {
-        config()->set('bootstrap-components.form.text.icon', null);
-        $html = bsText()->name('name')->toHtml();
+        config()->set('bootstrap-components.form.color.icon', null);
+        $html = bsColor()->name('name')->toHtml();
         $this->assertNotContains('<span class="icon input-group-text">', $html);
     }
 
     public function testHideIcon()
     {
         $configIcon = 'test-config-icon';
-        config()->set('bootstrap-components.form.text.icon', $configIcon);
-        $html = bsText()->name('name')->hideIcon()->toHtml();
+        config()->set('bootstrap-components.form.color.icon', $configIcon);
+        $html = bsColor()->name('name')->hideIcon()->toHtml();
         $this->assertNotContains('<span class="icon input-group-text">' . $configIcon . '</span>', $html);
     }
 
     public function testConfigLegend()
     {
         $configLegend = 'test-config-legend';
-        config()->set('bootstrap-components.form.text.legend', $configLegend);
-        $html = bsText()->name('name')->toHtml();
+        config()->set('bootstrap-components.form.color.legend', $configLegend);
+        $html = bsColor()->name('name')->toHtml();
         $this->assertContains(
-            '<small id="text-name-legend" class="form-text text-muted">bootstrap-components::'
+            '<small id="color-name-legend" class="form-text text-muted">bootstrap-components::'
             . $configLegend . '</small>',
             $html
         );
@@ -112,14 +112,14 @@ class TextTest extends BootstrapComponentsTestCase
     {
         $configLegend = 'test-config-legend';
         $customLegend = 'test-custom-legend';
-        config()->set('bootstrap-components.form.text.legend', $configLegend);
-        $html = bsText()->name('name')->legend($customLegend)->toHtml();
+        config()->set('bootstrap-components.form.color.legend', $configLegend);
+        $html = bsColor()->name('name')->legend($customLegend)->toHtml();
         $this->assertContains(
-            '<small id="text-name-legend" class="form-text text-muted">' . $customLegend . '</small>',
+            '<small id="color-name-legend" class="form-text text-muted">' . $customLegend . '</small>',
             $html
         );
         $this->assertNotContains(
-            '<small id="text-name-legend" class="form-text text-muted">bootstrap-components::'
+            '<small id="color-name-legend" class="form-text text-muted">bootstrap-components::'
             . $configLegend . '</small>',
             $html
         );
@@ -127,23 +127,23 @@ class TextTest extends BootstrapComponentsTestCase
 
     public function testNoLegend()
     {
-        config()->set('bootstrap-components.form.text.legend', null);
-        $html = bsText()->name('name')->toHtml();
-        $this->assertNotContains('<small id="text-name-legend" class="form-text text-muted">', $html);
+        config()->set('bootstrap-components.form.color.legend', null);
+        $html = bsColor()->name('name')->toHtml();
+        $this->assertNotContains('<small id="color-name-legend" class="form-text text-muted">', $html);
     }
 
     public function testHideLegend()
     {
         $configLegend = 'test-config-legend';
-        config()->set('bootstrap-components.form.text.legend', $configLegend);
-        $html = bsText()->name('name')->hideLegend()->toHtml();
-        $this->assertNotContains('<small id="text-name-legend" class="form-text text-muted">', $html);
+        config()->set('bootstrap-components.form.color.legend', $configLegend);
+        $html = bsColor()->name('name')->hideLegend()->toHtml();
+        $this->assertNotContains('<small id="color-name-legend" class="form-text text-muted">', $html);
     }
 
     public function testSetValue()
     {
         $customValue = 'test-custom-value';
-        $html = bsText()->name('name')->value($customValue)->toHtml();
+        $html = bsColor()->name('name')->value($customValue)->toHtml();
         $this->assertContains('value="' . $customValue . '"', $html);
     }
 
@@ -158,7 +158,7 @@ class TextTest extends BootstrapComponentsTestCase
             },
         ]);
         $this->call('GET', 'test');
-        $html = bsText()->name('name')->value($customValue)->toHtml();
+        $html = bsColor()->name('name')->value($customValue)->toHtml();
         $this->assertContains('value="' . $oldValue . '"', $html);
         $this->assertNotContains('value="' . $customValue . '"', $html);
     }
@@ -166,30 +166,30 @@ class TextTest extends BootstrapComponentsTestCase
     public function testSetLabel()
     {
         $label = 'test-custom-label';
-        $html = bsText()->name('name')->label($label)->toHtml();
-        $this->assertContains('<label for="text-name">' . $label . '</label>', $html);
+        $html = bsColor()->name('name')->label($label)->toHtml();
+        $this->assertContains('<label for="color-name">' . $label . '</label>', $html);
         $this->assertContains('placeholder="' . $label . '"', $html);
         $this->assertContains('aria-label="' . $label . '"', $html);
     }
 
     public function testNoLabel()
     {
-        $html = bsText()->name('name')->toHtml();
-        $this->assertContains('<label for="text-name">validation.attributes.name</label>', $html);
+        $html = bsColor()->name('name')->toHtml();
+        $this->assertContains('<label for="color-name">validation.attributes.name</label>', $html);
         $this->assertContains('aria-label="validation.attributes.name"', $html);
     }
 
     public function testHideLabel()
     {
-        $html = bsText()->name('name')->hideLabel()->toHtml();
-        $this->assertNotContains('<label for="text-name">validation.attributes.name</label>', $html);
+        $html = bsColor()->name('name')->hideLabel()->toHtml();
+        $this->assertNotContains('<label for="color-name">validation.attributes.name</label>', $html);
         $this->assertNotContains('aria-label="validation.attributes.name"', $html);
     }
 
     public function testSetPlaceholder()
     {
         $placeholder = 'test-custom-placeholder';
-        $html = bsText()->name('name')->placeholder($placeholder)->toHtml();
+        $html = bsColor()->name('name')->placeholder($placeholder)->toHtml();
         $this->assertContains('placeholder="' . $placeholder . '"', $html);
     }
 
@@ -197,20 +197,20 @@ class TextTest extends BootstrapComponentsTestCase
     {
         $label = 'test-custom-label';
         $placeholder = 'test-custom-placeholder';
-        $html = bsText()->name('name')->label($label)->placeholder($placeholder)->toHtml();
+        $html = bsColor()->name('name')->label($label)->placeholder($placeholder)->toHtml();
         $this->assertContains('placeholder="' . $placeholder . '"', $html);
     }
 
     public function testNoPlaceholder()
     {
-        $html = bsText()->name('name')->toHtml();
+        $html = bsColor()->name('name')->toHtml();
         $this->assertContains('placeholder="validation.attributes.name"', $html);
     }
 
     public function testSuccess()
     {
         $messageBag = app(MessageBag::class)->add('other_name', null);
-        $html = bsText()->name('name')->render(['errors' => $messageBag]);
+        $html = bsColor()->name('name')->render(['errors' => $messageBag]);
         $this->assertContains('<div class="valid-feedback d-block">', $html);
         $this->assertContains(
             trans('bootstrap-components::bootstrap-components.notification.validation.success'),
@@ -220,7 +220,7 @@ class TextTest extends BootstrapComponentsTestCase
 
     public function testNoSuccess()
     {
-        $html = bsText()->name('name')->toHtml();
+        $html = bsColor()->name('name')->toHtml();
         $this->assertNotContains('<div class="valid-feedback d-block">', $html);
     }
 
@@ -228,41 +228,41 @@ class TextTest extends BootstrapComponentsTestCase
     {
         $errorMessage = 'This a test error message';
         $messageBag = app(MessageBag::class)->add('name', $errorMessage);
-        $html = bsText()->name('name')->render(['errors' => $messageBag]);
+        $html = bsColor()->name('name')->render(['errors' => $messageBag]);
         $this->assertContains('<div class="invalid-feedback d-block">', $html);
         $this->assertContains($errorMessage, $html);
     }
 
     public function testNoError()
     {
-        $html = bsText()->name('name')->toHtml();
+        $html = bsColor()->name('name')->toHtml();
         $this->assertNotContains('<div class="invalid-feedback d-block">', $html);
     }
 
     public function testSetNoContainerId()
     {
-        $html = bsText()->name('name')->toHtml();
+        $html = bsColor()->name('name')->toHtml();
         $this->assertNotContains('<div id="', $html);
     }
 
     public function testSetContainerId()
     {
         $customContainerId = 'test-custom-container-id';
-        $html = bsText()->name('name')->containerId($customContainerId)->toHtml();
+        $html = bsColor()->name('name')->containerId($customContainerId)->toHtml();
         $this->assertContains('<div id="' . $customContainerId . '"', $html);
     }
 
     public function testSetNoComponentId()
     {
-        $html = bsText()->name('name')->toHtml();
-        $this->assertContains('for="text-name"', $html);
-        $this->assertContains('<input id="text-name"', $html);
+        $html = bsColor()->name('name')->toHtml();
+        $this->assertContains('for="color-name"', $html);
+        $this->assertContains('<input id="color-name"', $html);
     }
 
     public function testSetComponentId()
     {
         $customComponentId = 'test-custom-component-id';
-        $html = bsText()->name('name')->componentId($customComponentId)->toHtml();
+        $html = bsColor()->name('name')->componentId($customComponentId)->toHtml();
         $this->assertContains('for="' . $customComponentId . '"', $html);
         $this->assertContains('<input id="' . $customComponentId . '"', $html);
     }
@@ -270,44 +270,44 @@ class TextTest extends BootstrapComponentsTestCase
     public function testConfigContainerClass()
     {
         $configContainerCLass = 'test-config-class-container';
-        config()->set('bootstrap-components.form.text.class.container', [$configContainerCLass]);
-        $html = bsText()->name('name')->toHtml();
-        $this->assertContains('class="text-name-container ' . $configContainerCLass . '"', $html);
+        config()->set('bootstrap-components.form.color.class.container', [$configContainerCLass]);
+        $html = bsColor()->name('name')->toHtml();
+        $this->assertContains('class="color-name-container ' . $configContainerCLass . '"', $html);
     }
 
     public function testSetContainerClass()
     {
         $configContainerCLass = 'test-config-class-container';
         $customContainerCLass = 'test-custom-class-container';
-        config()->set('bootstrap-components.form.text.class.container', [$configContainerCLass]);
-        $html = bsText()->name('name')->containerClass([$customContainerCLass])->toHtml();
-        $this->assertContains('class="text-name-container ' . $customContainerCLass . '"', $html);
-        $this->assertNotContains('class="text-name-container ' . $configContainerCLass . '"', $html);
+        config()->set('bootstrap-components.form.color.class.container', [$configContainerCLass]);
+        $html = bsColor()->name('name')->containerClass([$customContainerCLass])->toHtml();
+        $this->assertContains('class="color-name-container ' . $customContainerCLass . '"', $html);
+        $this->assertNotContains('class="color-name-container ' . $configContainerCLass . '"', $html);
     }
 
     public function testConfigComponentClass()
     {
         $configComponentCLass = 'test-config-class-component';
-        config()->set('bootstrap-components.form.text.class.component', [$configComponentCLass]);
-        $html = bsText()->name('name')->toHtml();
-        $this->assertContains('class="form-control text-name-component ' . $configComponentCLass . '"', $html);
+        config()->set('bootstrap-components.form.color.class.component', [$configComponentCLass]);
+        $html = bsColor()->name('name')->toHtml();
+        $this->assertContains('class="form-control color-name-component ' . $configComponentCLass . '"', $html);
     }
 
     public function testSetComponentClass()
     {
         $configComponentCLass = 'test-config-class-component';
         $customComponentCLass = 'test-custom-class-component';
-        config()->set('bootstrap-components.form.text.class.component', [$customComponentCLass]);
-        $html = bsText()->name('name')->componentClass([$customComponentCLass])->toHtml();
-        $this->assertContains('class="form-control text-name-component ' . $customComponentCLass . '"', $html);
-        $this->assertNotContains('class="form-control text-name-component ' . $configComponentCLass . '"', $html);
+        config()->set('bootstrap-components.form.color.class.component', [$customComponentCLass]);
+        $html = bsColor()->name('name')->componentClass([$customComponentCLass])->toHtml();
+        $this->assertContains('class="form-control color-name-component ' . $customComponentCLass . '"', $html);
+        $this->assertNotContains('class="form-control color-name-component ' . $configComponentCLass . '"', $html);
     }
 
     public function testConfigContainerHtmlAttributes()
     {
         $configContainerAttributes = 'test-config-attributes-container';
-        config()->set('bootstrap-components.form.text.html_attributes.container', [$configContainerAttributes]);
-        $html = bsText()->name('name')->toHtml();
+        config()->set('bootstrap-components.form.color.html_attributes.container', [$configContainerAttributes]);
+        $html = bsColor()->name('name')->toHtml();
         $this->assertContains($configContainerAttributes, $html);
     }
 
@@ -315,8 +315,8 @@ class TextTest extends BootstrapComponentsTestCase
     {
         $configContainerAttributes = 'test-config-attributes-container';
         $customContainerAttributes = 'test-custom-attributes-container';
-        config()->set('bootstrap-components.form.text.html_attributes.container', [$configContainerAttributes]);
-        $html = bsText()->name('name')->containerHtmlAttributes([$customContainerAttributes])->toHtml();
+        config()->set('bootstrap-components.form.color.html_attributes.container', [$configContainerAttributes]);
+        $html = bsColor()->name('name')->containerHtmlAttributes([$customContainerAttributes])->toHtml();
         $this->assertContains($customContainerAttributes, $html);
         $this->assertNotContains($configContainerAttributes, $html);
     }
@@ -324,8 +324,8 @@ class TextTest extends BootstrapComponentsTestCase
     public function testConfigComponentHtmlAttributes()
     {
         $configComponentAttributes = 'test-config-attributes-component';
-        config()->set('bootstrap-components.form.text.html_attributes.component', [$configComponentAttributes]);
-        $html = bsText()->name('name')->toHtml();
+        config()->set('bootstrap-components.form.color.html_attributes.component', [$configComponentAttributes]);
+        $html = bsColor()->name('name')->toHtml();
         $this->assertContains($configComponentAttributes, $html);
     }
 
@@ -333,8 +333,8 @@ class TextTest extends BootstrapComponentsTestCase
     {
         $configComponentAttributes = 'test-config-attributes-component';
         $customComponentAttributes = 'test-custom-attributes-component';
-        config()->set('bootstrap-components.form.text.html_attributes.component', [$configComponentAttributes]);
-        $html = bsText()->name('name')->componentHtmlAttributes([$customComponentAttributes])->toHtml();
+        config()->set('bootstrap-components.form.color.html_attributes.component', [$configComponentAttributes]);
+        $html = bsColor()->name('name')->componentHtmlAttributes([$customComponentAttributes])->toHtml();
         $this->assertContains($customComponentAttributes, $html);
         $this->assertNotContains($configComponentAttributes, $html);
     }